Add files via upload
make request file and validation files

+++ app/Http/Controllers/BrandController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Requests\BrandValidation;
 use App\Models\Brand;
 use App\Models\Category;
 use Illuminate\Http\Request;
@@ -11,15 +12,15 @@ class BrandController extends Controller
     //
 
 
-public function store(Request $request){
-    $validaData = $request->validate([
+public function store(BrandValidation $brandValidation){
+   /* $validaData = $request->validate([
         'name' => 'required|max:255|min:3',
         'logoURL'=>'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
     ]);
     if(!$validaData){
         return response()->json(['error'=>'invalid request data']);
-    }
-    Category::create($validaData);
+    }*/
+    Category::create($brandValidation);
     return response()->json(['success'=>'Data added successfully.']);
 }
 public function show($id)
@@ -39,20 +40,26 @@ public function index()
 
 
 
-public function update(Request $request,$id){
-    $validaData = $request->validate([
+public function update(BrandValidation $brandValidation,$id){
+    /*$validaData = $request->validate([
         'name' => 'required|max:255|min:3',
         'logoURl'=>'required|image|mimes:jpeg,png,jpg,gif,svg|max:2048',
     ]);
     if(!$validaData){
         return response()->json(['error'=>'invalid request data']);
-    }
+    }*/
     $brand=Brand::find($id);
-    $brand->update($validaData);
+    if(!$brand){
+        return response()->json(['error'=>'brand not found']);
+    }
+    $brand->update($brandValidation);
     return response()->json(['success'=>'Data updated successfully.']);
 }
 public function destroy($id){
     $brand=Brand::find($id);
+    if(!$brand){
+        return response()->json(['error'=>'brand not found']);
+    }
     $brand->delete();
     return response()->json(['success'=>'Data deleted successfully.']);
 }
